pyproject.toml
[tool.poetry]
name = "embedding-search"
version = "0.1.0"
description = ""
authors = ["zawakin <zawawahoge@gmail.com>"]
readme = "README.md"
packages = [{include = "embedding_search"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.4"
pandas = "^2.0.1"
tiktoken = "^0.3.3"
matplotlib = "^3.7.1"
plotly = "^5.14.1"
scikit-learn = "^1.2.2"
pyqt5 = "^5.15.9"
ipywidgets = "^8.0.6"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
